+++ src/main/java/at/sparklingscience/urbantrees/domain/PhenologyObservationResult.java
@@ -19,7 +19,7 @@ public class PhenologyObservationResult {
 	 * Type identifier.
 	 */
 	@Min(1)
-	private int id;
+	private long id;
 	
 	/**
 	 * Type name
@@ -35,17 +35,17 @@ public class PhenologyObservationResult {
 	
 	public PhenologyObservationResult() {}
 	
-	public PhenologyObservationResult(int id, String description, int value) {
+	public PhenologyObservationResult(long id, String description, int value) {
 		this.setId(id);
 		this.setDescription(description);
 		this.setValue(value);
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
