+++ src/main/java/at/sparklingscience/urbantrees/domain/TreeSpecies.java
@@ -18,7 +18,7 @@ public class TreeSpecies implements Comparable<TreeSpecies> {
 	 * Database identifier.
 	 */
 	@Min(value = 1, groups = {ValidationGroups.Read.class})
-	private int id;
+	private long id;
 	
 	/**
 	 * Species name.
@@ -43,7 +43,7 @@ public class TreeSpecies implements Comparable<TreeSpecies> {
 	
 	@Override
 	public int compareTo(TreeSpecies o) {
-		int diff = o.getId() - this.getId();
+		long diff = o.getId() - this.getId();
 		if (diff > 0) {
 			return 1;
 		} else if (diff < 0) {
@@ -52,11 +52,11 @@ public class TreeSpecies implements Comparable<TreeSpecies> {
 		return 0;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
